[WIP] Really hacky way !!
Really HACKY way just to get a .wgt installed !!
this was just to check that we can do it this way,
basically by adding a LoadManifestWgt which should ideally
parse the config.xml and return a DictionaryValue, the same way
it is done for manifest.json.
We even keep the Manifest unchanged.
package, version, label/name, description, icon, are the only items
which are used by the PackageInstaller as of now.
So we need to parse only these fields.

Also added a enum in Package to distinguish between XPK and WGT

ISSUES :
in XPK, the package ID is generated from in the XPK c'stor using public key
In wgt the pacakgeId / application ID is in the wgt's config.xml file

thats why i disabled app_id check in application_service,
hence all packages are given the same garbage app_id.